fix(quick-open): score basename anchor so nested camelCase files match separator queries
