+++ src/renderer/src/components/quick-open-search.ts
@@ -1,8 +1,6 @@
 import { isClipboardTextByteLengthOverLimit } from '../../../shared/clipboard-text'
-import {
-  hasIdentifierTransitionMatchBeforeSeparator,
-  isPathSeparator
-} from './quick-open-word-boundaries'
+import { fuzzyMatchIndexedFile, prepareQuickOpenQuery } from './quick-open-fuzzy-match'
+import { isPathSeparator } from './quick-open-word-boundaries'
 
 export const QUICK_OPEN_RESULT_LIMIT = 50
 export const QUICK_OPEN_QUERY_MAX_BYTES = 2 * 1024
@@ -67,22 +65,11 @@ export function rankQuickOpenFiles(
   if (!normalizedQuery) {
     return files.slice(0, limit).map((file) => ({ path: file.path, score: 0 }))
   }
-  // Why: spaces, hyphens, and underscores are equivalent human separators, so
-  // camelCase basenames earn the filename boost for all three query styles.
-  const compactFilenameQuery = /[ _-]/.test(normalizedQuery)
-    ? normalizedQuery.replace(/[ _-]/g, '')
-    : null
-  const hasFlexibleIdentifierSeparator =
-    normalizedQuery.includes('_') || normalizedQuery.includes('-')
+  const preparedQuery = prepareQuickOpenQuery(normalizedQuery)
 
   const results: QuickOpenRankedResult[] = []
   for (const file of files) {
-    const score = fuzzyMatchIndexedFile(
-      normalizedQuery,
-      compactFilenameQuery,
-      hasFlexibleIdentifierSeparator,
-      file
-    )
+    const score = fuzzyMatchIndexedFile(preparedQuery, file)
     if (score === null) {
       continue
     }
@@ -91,196 +78,6 @@ export function rankQuickOpenFiles(
   }
 
   return results.map(({ path, score }) => ({ path, score }))
-}
-
-/**
- * Fuzzy subsequence match with VS Code-like word separators:
- * - Spaces in the query match path separators (`_`, `-`, `/`, `.`) or a
- *   zero-width camelCase / word boundary.
- * - Hyphens and underscores are interchangeable identifier separators.
- * - Word-start hits score better so basename/token starts rank above distant
- *   character soup.
- */
-function fuzzyMatchIndexedFile(
-  query: string,
-  compactFilenameQuery: string | null,
-  hasFlexibleIdentifierSeparator: boolean,
-  file: QuickOpenIndexedFile
-): number | null {
-  const path = file.lowerPath
-  const wordStarts = file.wordStarts
-  let qi = 0
-  let score = 0
-  let lastMatchIdx = -1
-  let requireWordStart = false
-
-  while (qi < query.length) {
-    if (query[qi] === ' ') {
-      qi++
-      // Trailing space should not occur after normalize, but stay safe.
-      if (qi >= query.length) {
-        break
-      }
-
-      const nextIdx = lastMatchIdx + 1
-      if (nextIdx < path.length && isPathSeparator(path[nextIdx])) {
-        // Separator runs are a single conceptual word break for spaced queries.
-        score -= 5
-        let sepIdx = nextIdx
-        while (sepIdx + 1 < path.length && isPathSeparator(path[sepIdx + 1])) {
-          sepIdx++
-        }
-        lastMatchIdx = sepIdx
-        requireWordStart = false
-      } else {
-        // Why: "Product Detail" vs ProductDetail — no separator char, only a
-        // camelCase boundary. Force the next letter onto a word start.
-        requireWordStart = true
-      }
-      continue
-    }
-
-    if (
-      isIdentifierSeparator(query[qi]) &&
-      qi + 1 < query.length &&
-      lastMatchIdx >= 0 &&
-      hasIdentifierTransitionMatchBeforeSeparator(path, wordStarts, lastMatchIdx + 1, query[qi + 1])
-    ) {
-      // Why: typed separators also bridge zero-width case transitions. Rank
-      // literal (0) < swapped separator (+1) < case transition (+2) so the
-      // order stays deterministic regardless of the filename boost.
-      score += 2
-      qi++
-      requireWordStart = true
-      continue
-    }
-
-    const wanted = query[qi]
-    let matched = false
-    for (let ti = lastMatchIdx + 1; ti < path.length; ti++) {
-      if (!searchCharactersMatch(path[ti], wanted)) {
-        continue
-      }
-      if (requireWordStart && wordStarts[ti] !== 1) {
-        continue
-      }
-
-      const gap = lastMatchIdx === -1 ? 0 : ti - lastMatchIdx - 1
-      score += gap
-      // Why: equivalent separators should match, but the literally typed form
-      // should rank first when both filename variants exist.
-      if (path[ti] !== wanted) {
-        score++
-      }
-      // Why: skip path index 0 so a leading `s` in `src/...` stays score 0
-      // (previous scorer only bonused matches after `/` `.` `-`).
-      if (ti > 0 && wordStarts[ti] === 1) {
-        score -= 5
-      }
-      lastMatchIdx = ti
-      qi++
-      requireWordStart = false
-      matched = true
-      break
-    }
-
-    if (!matched) {
-      // Why: a trailing separator mid-typing ("product-") must not blank out
-      // results that a trailing space would keep; literal was tried above.
-      if (isIdentifierSeparator(wanted) && qi === query.length - 1 && lastMatchIdx >= 0) {
-        score += 2
-        qi++
-        continue
-      }
-      return null
-    }
-  }
-
-  // Filename substring boost: allow human separators to hit equivalent names
-  // (`product detail` / `product-detail` → product_detail.dart).
-  if (
-    filenameMatchesQuery(
-      file.lowerFilename,
-      query,
-      compactFilenameQuery,
-      hasFlexibleIdentifierSeparator
-    )
-  ) {
-    score -= 100
-  }
-
-  return score
-}
-
-function filenameMatchesQuery(
-  lowerFilename: string,
-  query: string,
-  compactQuery: string | null,
-  hasFlexibleIdentifierSeparator: boolean
-): boolean {
-  if (
-    lowerFilename.includes(query) ||
-    (hasFlexibleIdentifierSeparator &&
-      includesWithIdentifierSeparatorEquivalence(lowerFilename, query))
-  ) {
-    return true
-  }
-  if (!compactQuery) {
-    return false
-  }
-  // Why: users type words with spaces; basenames use `_` / `-` / camelCase.
-  return includesIgnoringSeparators(lowerFilename, compactQuery)
-}
-
-function includesWithIdentifierSeparatorEquivalence(value: string, wanted: string): boolean {
-  const lastStart = value.length - wanted.length
-  for (let start = 0; start <= lastStart; start++) {
-    let offset = 0
-    while (offset < wanted.length && searchCharactersMatch(value[start + offset], wanted[offset])) {
-      offset++
-    }
-    if (offset === wanted.length) {
-      return true
-    }
-  }
-  return false
-}
-
-function includesIgnoringSeparators(value: string, wanted: string): boolean {
-  for (let start = 0; start < value.length; start++) {
-    if (value[start] !== wanted[0]) {
-      continue
-    }
-    let valueIndex = start
-    let wantedIndex = 0
-    while (valueIndex < value.length && wantedIndex < wanted.length) {
-      const ch = value[valueIndex]
-      if (isPathSeparator(ch)) {
-        valueIndex++
-        continue
-      }
-      if (ch !== wanted[wantedIndex]) {
-        break
-      }
-      valueIndex++
-      wantedIndex++
-    }
-    if (wantedIndex === wanted.length) {
-      return true
-    }
-  }
-  return false
-}
-
-function searchCharactersMatch(valueChar: string, queryChar: string): boolean {
-  return (
-    valueChar === queryChar ||
-    (isIdentifierSeparator(valueChar) && isIdentifierSeparator(queryChar))
-  )
-}
-
-function isIdentifierSeparator(ch: string): boolean {
-  return ch === '_' || ch === '-'
 }
 
 /**
